inline types used in one file
