--- packages/client/src/analytics/constants.ts
@@ -1,32 +1,0 @@
-// Split this up to avoid amplitude being pulled in when using these
-
-import { Context } from "decorator-shared/params";
-
-type AnalyticsCategory =
-    | "dekorator-header"
-    | "dekorator-footer"
-    | "dekorator-meny"
-    | "varsler";
-
-type AnalyticsActions =
-    | "søk-dynamisk"
-    | "navlogo"
-    | "lenke"
-    | "lenkegruppe"
-    | "hovedmeny/forsidelenke"
-    | "[redacted]"
-    | "nav.no"
-    | "arbeidsflate-valg"
-    | `${string}/${string}`
-    | string;
-
-export type AnalyticsEventArgs = {
-    eventName?: string;
-    category?: AnalyticsCategory;
-    action?: AnalyticsActions;
-    context?: Context;
-    destination?: string;
-    label?: string;
-    komponent?: string;
-    lenkegruppe?: "innlogget meny";
-};
